refactored handler test; fix test suite
